Rename `app-config` -> `sys-config`.
