+++ src/app-framework/export/BaseComponent.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { BaseConfig } from '@this/app-config';
+import { BaseConfig } from '@this/sys-config';
 import { IntfLogger } from '@this/loggy';
 import { MustBe } from '@this/typey';
