+++ src/app-framework/export/NetworkEndpoint.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { EndpointConfig, MountConfig } from '@this/app-config';
+import { EndpointConfig, MountConfig } from '@this/sys-config';
 import { TreePathKey, TreePathMap } from '@this/collections';
 import { IntfLogger } from '@this/loggy';
 import { DispatchInfo, IntfRateLimiter, IntfRequestHandler, IntfRequestLogger,
