+++ src/app-framework/export/Warehouse.js
@@ -3,7 +3,7 @@
 
 import * as timers from 'node:timers/promises';
 
-import { WarehouseConfig } from '@this/app-config';
+import { WarehouseConfig } from '@this/sys-config';
 import { PromiseUtil } from '@this/async';
 import { MustBe } from '@this/typey';
