+++ src/app-util/export/Saver.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { FileServiceConfig } from '@this/app-config';
+import { FileServiceConfig } from '@this/sys-config';
 import { IntfLogger } from '@this/loggy';
 
 import { BaseFilePreserver } from '#p/BaseFilePreserver';
