+++ src/app-util/package.json
@@ -12,7 +12,7 @@
   },
 
   "dependencies": {
-    "@this/app-config": "*",
+    "@this/sys-config": "*",
     "@this/async": "*",
     "@this/fs-util": "*",
     "@this/host": "*",
