+++ src/app-util/private/BaseFilePreserver.js
@@ -3,7 +3,7 @@
 
 import * as fs from 'node:fs/promises';
 
-import { FileServiceConfig } from '@this/app-config';
+import { FileServiceConfig } from '@this/sys-config';
 import { Condition, Threadlet } from '@this/async';
 import { Statter } from '@this/fs-util';
 import { IntfLogger } from '@this/loggy';
