+++ src/builtin-applications/package.json
@@ -12,7 +12,7 @@
   },
 
   "dependencies": {
-    "@this/app-config": "*",
+    "@this/sys-config": "*",
     "@this/app-framework": "*",
     "@this/fs-util": "*",
     "@this/loggy": "*",
