+++ src/builtin-services/package.json
@@ -12,7 +12,7 @@
   },
 
   "dependencies": {
-    "@this/app-config": "*",
+    "@this/sys-config": "*",
     "@this/app-framework": "*",
     "@this/app-util": "*",
     "@this/async": "*",
