+++ src/sys-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/app-config",
+  "name": "@this/sys-config",
   "version": "0.0.1",
   "type": "module",
   "private": true,
